removed locking mux for MoveFolder
